refactor: 그룹, 개별 내보내기 컴포넌트 통합 #20
