--- src/app/personal/[id]/export/Writer.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-import { content } from '@/types/content';
-
-interface Props {
-  content: content;
-}
-
-export default function Writer({ content }: Props) {
-  return (
-    <div className="inline">
-      {content.text}
-      <span className=" bg-emerald-400">{content.writer}</span>
-    </div>
-  );
-}
